Allow for recursive sourcing.
Do not allow the old python that is in the same directory as conda to
get set into the PATH which will override the python that is in the
conda env itself. This script should be able to be used as a way to
set up the PATH to both the conda executable AND the
conda-env-specific python, so that users can execute both from their
interactive shells.