Disallow adding $CONDA_INSTALL_DIR/bin to the PATH.
We simply CANNOT insert $CONDA_INSTALL_DIR/bin into the PATH at all in
the outer scope because if we did, and the user uses pip, then the pip
in the $CONDA_INSTALL_DIR/bin directory will be picked up and NOT the
pip that is under the conda-env where the python lives. If that
happens, then it messes the installation up in very weird ways.